pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "illuminate-mcp"
version = "0.1.0"
description = "Local MCP server for Anthology Illuminate on Snowflake"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Illuminate MCP Team" }
]
dependencies = []

[project.optional-dependencies]
snowflake = [
  "snowflake-connector-python==3.14.0"
]
dev = []

[project.scripts]
illuminate-mcp = "illuminate_mcp.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
